Return the last checked status from timer
Execting different loads to check for timer continuity and for returning
a result from the timer function doesn't necessarily mean the loads will
return the same results. On our code this will be always the case but
this make more sense since we want to get the conditions we checked for
during looping.