Continue pomodoro session if already started
If the pomodoro session has started by `timew continue` or if the hook
script failed to send a signal to `tw-pomodoro` pressing `(c)ontinue` in
the interface will automatically calculate the already tracked time and
complete the session with the remaining time.

If the already tracked time is greater than the configured session time
the session is ended and the break starts.

changelog:
- Timew class to manage the commands
- Fix bug when timew commands throw and isPause flag is not reseted
- Use `(c)ontinue` as command instead of `(s)tart`